requirements.txt
requests>=2.28.0
fuzzywuzzy[speedup]>=0.18.0
python-slugify>=8.0.0
PyPDF2>=3.0.0
jinja2>=3.1.0
bibtexparser>=1.2.0
python-docx >=0.8.11
markdown>=3.3.4
semantic-scholar-api #s2_folks    # If this is not available, use:scholar-scraper by Google Scholar
# claude mcp add semantic-scholar -- uvx semantic-scholar-mcp ## https://pypi.org/project/semantic-scholar-mcp/
textwrap3>=0.9.2
beautifulsoup4>=4.11.1
lxml>=4.9.1 
python-slugify>=8.0.0
openai>=0.27.0